Update package.json
If this builds on the Raspberry Pi (I don't currently have one), and I know you have limited space, but it doesn't look to be a large space addition when built; then it would build cleanly cross-platform. This works on linux and OSX.

`node-gyp 3.5.0` is -current, and is what I pulled on install, I have not tested earlier version.